Add Playwright BrowserInterface
