+++ index.js
@@ -1,4 +1,5 @@
 const BrowserInterfacePuppeteer = require( './lib/browser-interface-puppeteer' );
+const BrowserInterfacePlaywright = require( './lib/browser-interface-playwright' );
 const BrowserInterfaceIframe = require( './lib/browser-interface-iframe' );
 const generateCriticalCSS = require( './lib/generate-critical-css' );
 const { SuccessTargetError } = require( './lib/errors' );
@@ -7,6 +8,7 @@ module.exports = {
 	version: '0.0.1',
 	BrowserInterfaceIframe,
 	BrowserInterfacePuppeteer,
+	BrowserInterfacePlaywright,
 	generateCriticalCSS,
 	SuccessTargetError,
 };
