+++ lib/browser-interface-playwright.js
@@ -1,0 +1,11 @@
+const BrowserInterfaceNode = require( './browser-interface-node' );
+
+class BrowserInterfacePlaywright extends BrowserInterfaceNode {
+	constructor( pages ) {
+		super( pages );
+
+		this.interfaceName = 'Playwright';
+	}
+}
+
+module.exports = BrowserInterfacePlaywright;
