+++ lib/browser-interface-puppeteer.js
@@ -1,43 +1,10 @@
-const BrowserInterface = require( './browser-interface' );
+const BrowserInterfaceNode = require( './browser-interface-node' );
 
-class BrowserInterfacePuppeteer extends BrowserInterface {
+class BrowserInterfacePuppeteer extends BrowserInterfaceNode {
 	constructor( pages ) {
-		super();
+		super( pages );
 
-		this.pages = pages;
-	}
-
-	async runInPage( pageUrl, viewport, method, ...args ) {
-		const page = this.pages[ pageUrl ];
-
-		if ( ! page ) {
-			throw new Error(
-				`Puppeteer interface does not include URL ${ pageUrl }`
-			);
-		}
-
-		if ( viewport ) {
-			await page.setViewport( viewport );
-		}
-
-		// Get the inner window to pass to inner method.
-		const window = await page.evaluateHandle( () => window );
-
-		// Call inner method within the puppeteer context.
-		return page.evaluate( method, window, ...args );
-	}
-
-	/**
-	 * Replacement for browser.fetch, uses node-fetch to simulate the same
-	 * interface.
-	 *
-	 * @param {string} url URL to fetch.
-	 * @param {Object} options Fetch options.
-	 * @param {string} _role 'css' or 'html' indicating what kind of thing is being fetched.
-	 */
-	async fetch( url, options, _role ) {
-		const nodeFetch = require( 'node-fetch' );
-		return nodeFetch( url, options );
+		this.interfaceName = 'Puppeteer';
 	}
 }
